gles-user-module: fix issue that results in rc.pvr not being installed
The poky systemd class has a do_install() postfunc rm_sysvinit_initddir()
that deletes the /etc/init.d directory of a package if the systemd
system unit directory contains some files. gles-user-module supports
both systemd and sysv but uses /etc/init.d in both cases as the location
of the required script rc.pvr. In the case when systemd only is enabled
the class deletes the init.d directory and the file.

Until the script is converted to a service file workaround the issue
by moving the install location of rc.pvr when systemd alone is used.